refactor(consumer): extract session into its own file
Move the ConsumerGroupSession, ConsumerGroupHandler and
ConsumerGroupClaim types and their methods out of consumer_group.go
into consumer_group_session.go with no behavior change, to make room
for the cooperative rebalancing lifecycle.
